Break out tokio test, use required flavor=multi_thread.
