+++ tests/integration_test.rs
@@ -33,17 +33,19 @@ fn test_eval() {
     assert!(ui.eval("dtyfhgxnt*").is_err());
 }
 
-#[tokio::test]
-async fn test_bind_async() {
+#[test]
+fn test_bind_async() {
     let ui = UIBuilder::new()
-        .content(Content::Html(r#"
+        .content(Content::Html(
+            r#"
         <script>
         async function foo(x) {
           const result = await bar(x + 'b');
           return result + 'd';
         }
         </script>
-        "#))
+        "#,
+        ))
         .custom_args(&["--headless"])
         .run()
         .expect("Unable to launch");
@@ -54,32 +56,10 @@ async fn test_bind_async() {
             .spawn(move || {
                 let result = format!("{}c", context.args()[0].as_str().expect("arg to be str"));
                 context.complete(Ok(result.into()))
-            });
-    }).unwrap();
-
-    assert_eq!(ui.eval("foo('a')").unwrap(), "abcd");
-}
-
-#[cfg(feature = "tokio")]
-#[tokio::test]
-async fn test_bind_tokio() {
-    let ui = UIBuilder::new()
-        .content(Content::Html(r#"
-        <script>
-        async function foo(x) {
-          const result = await bar(x + 'b');
-          return result + 'd';
-        }
-        </script>
-        "#))
-        .custom_args(&["--headless"])
-        .run()
-        .expect("Unable to launch");
-
-    ui.bind_tokio("bar", move |args| async move {
-        tokio::task::yield_now().await;
-        Ok(format!("{}c", args[0].as_str().expect("arg to be str")).into())
-    }).unwrap();
+            })
+            .expect("failed to spawn thread");
+    })
+    .unwrap();
 
     assert_eq!(ui.eval("foo('a')").unwrap(), "abcd");
 }
